Consolidate sys.excepthook replacements
